Implement Evaluators.
A Rule can hold an evaluator, which will evaluate to an Object, given
the result of parsing (i.e. a ParsedNode).

E.g. one can define a rule for an integer, and attach a custom evaluator
which creates an int from the parsed string, which is obtained from the
ParsedNode.