Fix VETO works for autocompletion.
VETO is used to prevent autocompletion in case one of the
autocompletions is a parameterized autocompletion and the user already
started to type (in which case no further autocompletion should be
performed).

E.g. when autocompleting an integer, autocompletion would first insert
the name of the integer, but once the user started to type digits, no
further autocompletion should take place (in particular, the name of the
integer should not be inserted once more).