Major commit, fixing primarily autocompletion.
Everything used to work fine as long as parsing was successful, but the
generated parse tree was not correct in case of autocompletion, in fact,
there are individual (and different) parse tree for each encountered
end-of-input parsing possibility.

Parsing and parse tree generation are now separated. Parsing itself
generates a tree of SymbolSequences.